Full implementation on the unit test coverage - broke the flash simulator out of unit test main and into it's own c file. Able to flash an image into the simulated flash memory and boot from it.
