+++ src/bl_version.h
@@ -3,6 +3,6 @@
 
 #define BL_VERSION_MAJOR 0
 #define BL_VERSION_MINOR 0
-#define BL_VERSION_REV 15
+#define BL_VERSION_REV 16
 
 #endif /* BL_VERSION_H */
