+++ utilities/version/version.json
@@ -1,6 +1,6 @@
 {
   "BL_VERSION_MAJOR": 0,
   "BL_VERSION_MINOR": 0,
-  "BL_VERSION_REV": 15,
+  "BL_VERSION_REV": 16,
   "EXT_FLASH_VERSION": "1.0.21"
 }
